requirements.txt
django
djangorestframework
drf-writable-nested